% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_file.R
\name{dalocon_download_xls}
\alias{dalocon_download_xls}
\title{Download excel files easily}
\usage{
dalocon_download_xls(
  appID,
  urls,
  dest_dir = "./",
  check = TRUE,
  make_dir = FALSE
)
}
\arguments{
\item{appID}{e-Stat API's application id}

\item{urls}{excel files urls}

\item{dest_dir}{directory which you want to put files}

\item{check}{if True, duplicate file is not download.}

\item{make_dir}{if True, force to create directory, if not exist}
}
\description{
Download excel files easily
}
