% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls_maker.R
\name{dalocon_get_table15_urls}
\alias{dalocon_get_table15_urls}
\title{get urls for table nunber 15}
\usage{
dalocon_get_table15_urls(appID)
}
\arguments{
\item{appID}{e-Stat API's application id}
}
\value{
urls vector
}
\description{
get urls for table nunber 15
}
