% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_t07.R
\name{dalocon_make_t07_tidy}
\alias{dalocon_make_t07_tidy}
\title{Make table number 07 data tidy}
\usage{
dalocon_make_t07_tidy(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
data.frame
}
\description{
Make table number 07 data tidy
}
