% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_t07.R
\name{dalocon_read_t07_xls}
\alias{dalocon_read_t07_xls}
\title{Read xls files of table number 7}
\usage{
dalocon_read_t07_xls(files)
}
\arguments{
\item{files}{xls file's path vector}
}
\value{
data.frame
}
\description{
Read xls files of table number 7
}
