% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_t15.R
\name{dalocon_read_t15_xls}
\alias{dalocon_read_t15_xls}
\title{Read xls files of table number 15}
\usage{
dalocon_read_t15_xls(files)
}
\arguments{
\item{files}{xls file's path vector}
}
\value{
data.frame
}
\description{
Read xls files of table number 15
}
