% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls_maker.R
\name{sub_get_metadata}
\alias{sub_get_metadata}
\title{Get metadata}
\usage{
sub_get_metadata(appID, tablenum = "15")
}
\arguments{
\item{appID}{estat apprication id}

\item{tablenum}{table number}
}
\value{
df data.frame
}
\description{
Get metadata
}
